Project_APP/backend/requirements.txt
# Project Dependencies
django==5.0.2
djangorestframework==3.14.0
django-apscheduler==0.6.2
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2
apscheduler==3.10.4
python-dateutil==2.8.2